updated readme.md and docker compose fix
