--- backend/README.md
@@ -1,5 +1,0 @@
-# Setup
-use Venv, Conda or system python and install the requirements.txt
-
-# Run
-python main.py
