--- backend/requirements.txt
@@ -1,18 +1,0 @@
-fastapi>=0.115.8
-uvicorn[standard]>=0.34.0
-langchain>=0.3.18
-langchain_community>=0.3.17
-openai>=1.63.0
-psycopg[binary]>=3.2.4
-psycopg[pool]>=3.2.4
-langgraph>=0.2.73
-langchain_openai>=0.3.12
-azure-storage-blob
-python-multipart
-shapely
-kml2geojson
-geopandas
-webcolors
-pytest-asyncio
-pytest
-
