+++ frontend/app/components/chat/AgentInterface.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client";
 
 import { useState, useEffect, useRef } from "react";
 import wellknown from "wellknown";
