+++ frontend/app/components/maps/MapComponent.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client";
 
 import LeafletMapComponent from "./LeafletMap";
