+++ frontend/app/components/sidebar/LayerManagement.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client";
 
 import { useRef, useState, useEffect } from "react";
 import { useMapStore } from "../../stores/mapStore";
