+++ frontend/app/components/sidebar/Sidebar.tsx
@@ -1,4 +1,4 @@
-'use client'
+"use client"
 
 import { useState } from 'react';
 import Link from 'next/link';
